+++ lib/common/Makefile.am
@@ -47,10 +47,13 @@ endif
 libcrmcommon_la_SOURCES	+= cmdline.c
 libcrmcommon_la_SOURCES	+= digest.c
 libcrmcommon_la_SOURCES	+= io.c
-libcrmcommon_la_SOURCES	+= ipc.c
+libcrmcommon_la_SOURCES	+= ipc_client.c
+libcrmcommon_la_SOURCES	+= ipc_common.c
+libcrmcommon_la_SOURCES	+= ipc_server.c
 libcrmcommon_la_SOURCES	+= iso8601.c
 libcrmcommon_la_SOURCES	+= logging.c
 libcrmcommon_la_SOURCES	+= mainloop.c
+libcrmcommon_la_SOURCES	+= messages.c
 libcrmcommon_la_SOURCES	+= nvpair.c
 libcrmcommon_la_SOURCES	+= operations.c
 libcrmcommon_la_SOURCES	+= options.c
